requirements.txt
PyMySQL==1.1.1
python-dotenv==1.1.0
sniffio==1.3.1
typing_extensions==4.13.0
fastapi
uvicorn
pydantic
sqlalchemy
databases